[Add] Odin Inspector 에셋 추가

+++ Assets/02.Scripts/YoYo/YoYoData.cs
@@ -8,9 +8,9 @@ public class YoYoData : InfiniteScrollData
     public int id; // id
     public string name; // 이름
     public string type; // 타입 (디스크, 와이어)
-    public int attack; // 공격력
-    public int attackSpeed; // 공격 속도
-    public int attackRange; // 공격 사거리
+    public float attack; // 공격력
+    public float attackSpeed; // 공격 속도
+    public float attackRange; // 공격 사거리
     public string imagePath; // 이미지 경로
 }
